feat(amd): HIP 35B — bf16-attn route + prefill __half + reshape guard
Toward native-HIP qwen3.6-35b-a3b. Drop the FP8-native attention kernels
(w8a16_gemm[_t/_pipelined], fp8_gemm_t_blockscaled) from kernels/strix-hip —
hipcc can't compile their inline PTX (mma.sync/cp.async) and they lack WMMA
copies. On HIP the 35B routes attention through BF16 via the existing
ATLAS_FP8_DEQUANT_ATTN_TO_BF16 lever → dense_gemm_bf16 (which has a WMMA copy),
so these kernels aren't needed.

Kernel fixes (subagent-assisted, verified):
- prefill_paged_compute.cuh: __half→__fp16 PV-WMMA casts routed through
  __half2float (lines 329/599) — value-identical, hipcc-clean.
- reshape_and_cache_turbo.cu: the cvt.rn.satfinite.e4m3x2.f32 `=h` PTX guarded
  → scl_enc_fp8 on SCALE/HIP (bit-exact SATFINITE E4M3), NVIDIA asm verbatim
  in #else.

NEXT HIP blocker: cp.async in the LOAD_KV_TILE macro (inferspark_prefill_paged*
.cu) — replace with synchronous uint4 copy (its #else-zero branch already is
one) via a guarded atlas_cp16 helper; prune unused prefill variants from
strix-hip. NVIDIA unaffected (strix-hip is HIP-only; reshape #else verbatim).

+++ kernels/strix-hip/common/prefill_paged_compute.cuh
@@ -326,7 +326,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16h a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__fp16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__fp16)__half2float(smem_P[pv_warp_m + lane_lo][k_off + i]);
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -596,7 +596,7 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16h a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__fp16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__fp16)__half2float(smem_P[pv_warp_m + lane_lo][k_off + i]);
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
                     unsigned int d_col = (pv_n_start + nt) * 16 + lane_lo;
